Disable confirm buttons after confirm operation
